refactored Patching class
